some fix and format

+++ src/routes/post/[id]/edit/+page.server.ts
@@ -1,25 +1,26 @@
 import type { PageServerLoad, Actions } from './$types';
 import { posts, contents } from '$lib/schema';
-import { and, eq } from 'drizzle-orm';
+import { and, eq, sql } from 'drizzle-orm';
 import { error, redirect } from '@sveltejs/kit';
 
-export const load: PageServerLoad = async ({ params, parent, locals }) => {
-	const id = params.id;
-	const { session } = await parent();
-	if (!session || !session.user) throw redirect(302, '/signin');
+export const load: PageServerLoad = async ({ params, locals }) => {
+	const session = await locals.auth.validate();
+	if (!session) throw error(302, '/signin');
+	if (!session.user.emailVerified) throw redirect(302, '/email-verification');
 	const post = await locals.db
 		.select({
+			id: posts.id,
 			title: posts.title,
 			description: contents.description,
 			thumbnail: contents.thumbnail,
 			file: contents.file
 		})
 		.from(posts)
-		.where(and(eq(posts.id, id), eq(posts.authorId, session.user.userId)))
+		.where(and(eq(posts.id, params.id), eq(posts.authorId, session.user.userId)))
 		.leftJoin(contents, and(eq(posts.version, contents.version), eq(posts.id, contents.post)))
 		.get();
 	if (!post) throw error(404, 'Post not found or you are not the author');
-	return { post, id };
+	return { post };
 };
 
 export const actions: Actions = {
@@ -27,15 +28,13 @@ export const actions: Actions = {
 		const session = await locals.auth.validate();
 		if (!session) throw error(401, 'Unauthorized');
 		if (!session.user.emailVerified) throw redirect(302, '/email-verification');
+
 		const post = await locals.db
-			.select({
-				authorId: posts.authorId
-			})
+			.select({ count: sql<number>`COUNT(*)` })
 			.from(posts)
-			.where(eq(posts.id, params.id))
+			.where(and(eq(posts.id, params.id), eq(posts.authorId, session.user.userId)))
 			.get();
-		if (!post) throw error(404, 'Post not found');
-		if (post.authorId != session.user.userId) throw error(403, 'Forbidden');
+		if (!post) throw error(404, 'Post not found or you are not the author');
 		await locals.db.delete(posts).where(eq(posts.id, params.id)).execute();
 		await locals.db.delete(contents).where(eq(contents.post, params.id)).execute();
 		return {
@@ -57,11 +56,10 @@ export const actions: Actions = {
 				file: contents.file
 			})
 			.from(posts)
-			.where(eq(posts.id, params.id))
-			.leftJoin(contents, and(eq(posts.version, contents.version), eq(posts.id, contents.post)))
+			.where(and(eq(posts.id, params.id), eq(posts.authorId, session.user.userId)))
+			.innerJoin(contents, and(eq(posts.version, contents.version), eq(posts.id, contents.post)))
 			.get();
-		if (!post) throw error(404, 'Post not found');
-		if (post.authorId != session.user.userId) throw error(403, 'Forbidden');
+		if (!post) throw error(404, 'Post not found or you are not the author');
 		const formData = Object.fromEntries(await request.formData());
 		const differences: {
 			title?: string;
